+++ CHANGELOG.rst
@@ -7,7 +7,7 @@ Versions are year-based with a strict backward compatibility policy.
 The third digit is only for regressions.
 
 
-18.3.0 (UNRELEASED)
+18.3.0 (2018-06-21)
 -------------------